Change axis to display only integer values

+++ src/main/java/seedu/address/model/ModelManager.java
@@ -176,9 +176,9 @@ public class ModelManager implements Model {
     @Override
     public void updatePerformance(Model model) {
         requireNonNull(model);
-        int numGood = model.getFilteredFlashCardListNoCommit(new RatingContainsKeywordPredicate("good")).size();
-        int numHard = model.getFilteredFlashCardListNoCommit(new RatingContainsKeywordPredicate("hard")).size();
-        int numEasy = model.getFilteredFlashCardListNoCommit(new RatingContainsKeywordPredicate("easy")).size();
+        float numGood = model.getFilteredFlashCardListNoCommit(new RatingContainsKeywordPredicate("good")).size();
+        float numHard = model.getFilteredFlashCardListNoCommit(new RatingContainsKeywordPredicate("hard")).size();
+        float numEasy = model.getFilteredFlashCardListNoCommit(new RatingContainsKeywordPredicate("easy")).size();
         float value;
         if ((numEasy + numGood + numHard) == 0) {
             value = 0;
